projectmain/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-cors-headers==3.13.0
joblib==1.2.0
numpy==1.24.1
Pillow==9.4.0
scikit-learn==1.2.0
scipy==1.10.0
sqlparse==0.4.3
threadpoolctl==3.1.0
tzdata==2022.7
whitenoise==6.3.0
